✨ feat: Moved modal components from src/modal to src/components/modal.
