+++ src/components/modal/EmbeddingConfig.tsx
